setting up profile page
